+++ frontend/src/components/Carta/Revelaciones.module.css
@@ -1,0 +1,25 @@
+.modalDiv {
+    z-index: 3;
+}
+
+.modalWindow {
+    text-align: center;
+    color: white;
+    position: relative;
+    top: 25%;
+}
+
+.modalBody {
+    background-color: #121212;
+}
+
+.modalBody h4 {
+    text-align: center;
+    padding: 35px;
+}
+
+.divBotonModal {
+    display: flex;
+    justify-content: space-evenly;
+    margin: 10px 0;
+}
