+++ frontend/src/components/Partida/Partida.module.css
@@ -33,29 +33,3 @@
     width: 80px;
     border-radius: 10px;
 }
-
-.modalDiv {
-    z-index: 3;
-}
-
-.modalWindow {
-    text-align: center;
-    color: white;
-    position: relative;
-    top: 25%;
-}
-
-.modalBody {
-    background-color: #121212;
-}
-
-.modalBody h4 {
-    text-align: center;
-    padding: 35px;
-}
-
-.divBotonModal {
-    display: flex;
-    justify-content: space-evenly;
-    margin: 10px 0;
-}
